Исправляет замечания наставника
