+++ index.php
@@ -1,6 +1,7 @@
 <?php
 require_once('helpers.php');
 require_once('init.php');
+require_once('functions.php');
 
 $is_auth = rand(0, 1);
 
@@ -28,22 +29,6 @@ $result_categories = handle_query($connect, $sql_categories);
 $categories = mysqli_fetch_all($result_categories, MYSQLI_ASSOC);
 $goods = mysqli_fetch_all($result_lots, MYSQLI_ASSOC);
 
-/**
- * @param float $old_price
- *
- * @author Trikashnyi Artem tema-luch@mail.ru
- *
- * @return string
- */
-
-function format_price(float $old_price) {
-    $new_price = ceil($old_price);
-
-    $new_price = $new_price < 1000 ? $new_price . ' ₽' : number_format($new_price, 0, '', ' ') . ' ₽';
-
-    return $new_price;
-}
-
 $page_content = include_template('main.php',
     [
         'categories' => $categories,
@@ -60,21 +45,5 @@ $layout_content = include_template('layout.php',
         'categories' => $categories
     ]
 );
-
-/**
- * @param string $date
- *
- * @author Trikashnyi Artem tema-luch@mail.ru
- *
- * @return int[]
- */
-
-function get_time_range(string $date) {
-    $time_difference = strtotime($date) - time();
-    $hours_count = floor($time_difference / 3600);
-    $minutes_count = floor(($time_difference % 3600) / 60);
-
-    return [$hours_count, $minutes_count];
-}
 
 print($layout_content);
